DevOp/Doсkers/dz_1_2/3.2-crud/stocks_products/requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
django-filter
gunicorn